requirements.txt
flask
pandas
sqlalchemy
python-dotenv
openai
requests